% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cpnor.R
\name{Cpnor}
\alias{Cpnor}
\title{the empirical coverage rates for the sample mean}
\usage{
Cpnor(theta, n, m)
}
\arguments{
\item{theta}{the skewness of normal distribution}

\item{n}{the size of each sample you want}

\item{m}{the times of iteration you want}
}
\value{
a coverage probabilities between 0 and 1
}
\description{
a Monte Carlo study to estimate the coverage probabilities of the standard normal bootstrap confidence interval
}
\examples{
n <- 20; m <- 1000
Cpnor(0,n,m)
}
