% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomwalkC}
\alias{randomwalkC}
\title{Random Walk using Rcpp}
\usage{
randomwalkC(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of the normal random increment with a zero mean}

\item{x0}{the initial iteration point}

\item{N}{the designated number of random numbers (including the initial point x0)}
}
\value{
the sample
}
\description{
A Random Walk sampler using Rcpp
}
\examples{
\dontrun{
laplace<-function(x) return(1/2*exp(-abs(x)))
lC <- randomwalkC(1,25,1000);
}
}
