% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rl.R
\name{Rl}
\alias{Rl}
\title{Generate samples from Rayleigh distribution}
\usage{
Rl(n, sigma)
}
\arguments{
\item{n}{the sample size}

\item{sigma}{the parameter of Rayleigh distribution}
}
\value{
a random sample of size n
}
\description{
Generate samples from Rayleigh distribution
}
\examples{
n <- 100
sample <- Rayleigh(n,2)
head(sample,10)
}
