% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwme}
\alias{rwme}
\title{Random Walk using Rcpp}
\usage{
rwme(sigma, xo, N)
}
\arguments{
\item{sigma}{the standard deviation of the normal random increment with a zero mean}

\item{N}{the designated number of random numbers (including the initial point x0)}

\item{x0}{the initial iteration point}
}
\value{
the sample
}
\description{
A Random Walk sampler using Rcpp
}
\examples{
\dontrun{
laplace<-function(x) return(1/2*exp(-abs(x)))
lC <- rwme(1,25,1000);
plot(1:1000,lC,type='l')
}
}
