\name{SC19013-package}
\alias{SC19013-package}
\alias{SC19013}
\docType{package}
\title{
  A simple R package contain two fuctions and my whole homework.
}
\description{
  The first function called Rl is to generate samples from  Rayleigh distribution by using antithetic variables. The second function called Cpnor is a Monte Carlo study to estimate the coverage probabilities of the standard normal bootstrap confidence interval.
}

\author{
Honghong Xu.

Maintainer: Honghong Xu <hhongxu@mail.ustc.edu.cn>
}
\keyword{ package }
